Fixed issue with 'when' property of prompts not working on first question

+++ remove/index.js
@@ -16,36 +16,44 @@ module.exports = yeoman.generators.NamedBase.extend({
   askFor: function () {
     var done = this.async();
 
-    var moduleExists = function (rootPath) {
+    var getModulePath = function (rootPath) {
       rootPath = rootPath || 'app';
-      this.modulePath = path.join(this.env.cwd, 'src', rootPath, this.name);
-      return fs.existsSync(this.modulePath);
+      return path.join(this.env.cwd, 'src', rootPath, this.name);
     }.bind(this);
 
-    var prompts = [
-      {
+    var moduleExists = function (rootPath) {
+      return fs.existsSync(getModulePath(rootPath));
+    };
+
+    var prompts = [];
+
+    if (!moduleExists()) {
+      prompts.push({
         name: 'modulePath',
         message: 'Where is the location of this module?',
         default: 'app',
-        when: !moduleExists,
         validate: moduleExists
-      }, {
-        name: 'remove',
-        message: 'Are you sure you want to remove ' + this.name + '?',
-        type: 'confirm'
-      }
-    ];
+      });
+    }
+
+    prompts.push({
+      name: 'remove',
+      message: 'Are you sure you want to remove ' + this.name + '?',
+      type: 'confirm'
+    });
 
     this.prompt(prompts, function (props) {
-      moduleExists(props.modulePath);
+      this.modulePath = getModulePath(props.modulePath);
       this.removeModule = props.remove;
       done();
     }.bind(this));
   },
 
   files: function () {
-    fs.removeSync(this.modulePath);
-    this._updateAppJs(this.name);
+    if (this.removeModule) {
+      fs.removeSync(this.modulePath);
+      this._updateAppJs(this.name);
+    }
   },
 
   _updateAppJs: function (camelModuleName) {
